+++ yggdrasil_decision_forests/port/python/ydf/learner/learner_test.py
@@ -15,6 +15,7 @@
 """Tests for model learning."""
 
 import os
+import random
 import signal
 from typing import Any, Dict, Optional, Tuple
 
@@ -25,7 +26,7 @@ import fastavro
 import numpy as np
 import numpy.testing as npt
 import pandas as pd
-import polars as pl
+# import polars as pl # TODO: Re-enable.
 from sklearn import metrics
 
 from yggdrasil_decision_forests.dataset import data_spec_pb2 as ds_pb
@@ -437,7 +438,7 @@ class RandomForestLearnerTest(LearnerTest):
         self.two_center_regression.train, folds=10, parallel_evaluations=2
     )
     logging.info("evaluation:\n%s", evaluation)
-    self.assertAlmostEqual(evaluation.rmse, 116, delta=1)
+    self.assertAlmostEqual(evaluation.rmse, 118, delta=3)
     # All the examples are used in the evaluation
     self.assertEqual(
         evaluation.num_examples,
@@ -971,6 +972,35 @@ class RandomForestLearnerTest(LearnerTest):
           label="label", features=[("f1", dataspec.Semantic.HASH)], num_trees=2
       ).train(data)
 
+  def test_numerical_vector_sequence(self):
+    def make_ds():
+      features = []
+      labels = []
+      num_examples = 1_000
+      num_dims = 2
+      distance_limit = 0.2
+      for _ in range(num_examples):
+        num_vectors = random.randint(0, 5)
+        vectors = np.random.uniform(
+            size=(num_vectors, num_dims), low=0.0, high=1.0
+        )
+        label = np.any(
+            np.sum((vectors - np.array([[0.5, 0.5]])) ** 2, axis=1)
+            < distance_limit * distance_limit
+        )
+        features.append(vectors)
+        labels.append(label)
+      return {"label": np.array(labels), "features": features}
+
+    train_ds = make_ds()
+    test_ds = make_ds()
+    model = specialized_learners.RandomForestLearner(
+        label="label", num_trees=50
+    ).train(train_ds)
+    evaluation = model.evaluate(test_ds)
+    logging.info("Evaluation: %s", evaluation)
+    self.assertGreaterEqual(evaluation.accuracy, 0.95)
+
 
 class CARTLearnerTest(LearnerTest):
 
@@ -1001,7 +1031,7 @@ class CARTLearnerTest(LearnerTest):
     )
     model = learner.train(self.two_center_regression.train)
     evaluation = model.evaluate(self.two_center_regression.test)
-    self.assertAlmostEqual(evaluation.rmse, 114.081, places=3)
+    self.assertAlmostEqual(evaluation.rmse, 116, delta=0.5)
 
   def test_monotonic_non_compatible_learner(self):
     with self.assertRaisesRegex(
@@ -1481,7 +1511,7 @@ class GradientBoostedTreesLearnerTest(LearnerTest):
     ).train(data)
     npt.assert_equal(model_1.label_classes(), np.unique(label_data).astype(str))
 
-  def test_adult_poison(self):
+  def test_adult_poisson(self):
     model = specialized_learners.GradientBoostedTreesLearner(
         label="hours_per_week",
         growing_strategy="BEST_FIRST_GLOBAL",
@@ -1531,7 +1561,11 @@ class IsolationForestLearnerTest(LearnerTest):
     _ = model.describe("text")
     _ = model.describe("html")
     _ = model.analyze_prediction(self.gaussians.test_pd.iloc[:1])
-    _ = model.analyze(self.gaussians.test_pd)
+    analysis = model.analyze(self.gaussians.test_pd)
+    if with_labels:
+      self.assertLen(analysis.variable_importances(), 7)
+    else:
+      self.assertLen(analysis.variable_importances(), 3)
 
   def test_gaussians_evaluation_default_task(self):
     learner = specialized_learners.IsolationForestLearner(label="label")
@@ -1660,41 +1694,49 @@ class DatasetFormatsTest(parameterized.TestCase):
         "multi_f1",
     ]
 
-  def create_polars_dataset(self, n: int = 1000) -> pl.DataFrame:
-    return pl.DataFrame({
-        # Single-dim features
-        "f1": np.random.random(size=n),
-        "f2": np.random.random(size=n),
-        "i1": np.random.randint(100, size=n),
-        "i2": np.random.randint(100, size=n),
-        "c1": np.random.choice(["x", "y", "z"], size=n, p=[0.6, 0.3, 0.1]),
-        "multi_c1": np.array(
-            [["a", "x", "z"], ["b", "x", "w"], ["a", "y", "w"], ["b", "y", "z"]]
-            * (n // 4)
-        ),
-        # Cat-set features
-        # ================
-        # Note: Polars as a bug when serializing empty lists of string to Avro
-        # files (only write one of the two required "optional" bit).
-        # TODO: Replace [""] by [] once the bug if fixed is added.
-        "cs1": [["<SOMETHING>"], ["a", "b", "c"], ["b", "c"], ["a"]] * (n // 4),
-        # Multi-dim features
-        # ==================
-        # Note: It seems support for this type of feature was temporarly dropped
-        # in Polars 1.9 i.e. the data packing was improved but the avro
-        # serialization was not implemented. This code would fail with recent
-        # version of polars with: not yet implemented: write
-        # FixedSizeList(Field { name: "item", dtype: Float64, is_nullable: true,
-        # metadata: {} }, 5) to avro.
-        "multi_f1": np.random.random(size=(n, 3)),
-        # # Labels
-        "label_class_binary1": np.random.choice([False, True], size=n),
-        "label_class_binary2": np.random.choice([0, 1], size=n),
-        "label_class_binary3": np.random.choice(["l1", "l2"], size=n),
-        "label_class_multi1": np.random.choice(["l1", "l2", "l3"], size=n),
-        "label_class_multi2": np.random.choice([0, 1, 2], size=n),
-        "label_regress1": np.random.random(size=n),
-    })
+  # TODO: Re-enable.
+  def create_polars_dataset(self, n: int = 1000):
+    del n
+    raise ValueError("Not available")
+
+  # def create_polars_dataset(self, n: int = 1000) -> pl.DataFrame:
+  #   return pl.DataFrame({
+  #       # Single-dim features
+  #       "f1": np.random.random(size=n),
+  #       "f2": np.random.random(size=n),
+  #       "i1": np.random.randint(100, size=n),
+  #       "i2": np.random.randint(100, size=n),
+  #       "c1": np.random.choice(["x", "y", "z"], size=n, p=[0.6, 0.3, 0.1]),
+  #       "multi_c1": np.array(
+  #           [["a", "x", "z"], ["b", "x", "w"], ["a", "y", "w"],
+  #           ["b", "y", "z"]]
+  #           * (n // 4)
+  #       ),
+  #       # Cat-set features
+  #       # ================
+  #       # Note: Polars as a bug when serializing empty lists of string to Avro
+  #       # files (only write one of the two required "optional" bit).
+  #       # TODO: Replace [""] by [] once the bug if fixed is added.
+  #       "cs1": [["<SOMETHING>"], ["a", "b", "c"], ["b", "c"], ["a"]]
+  #         * (n // 4),
+  #       # Multi-dim features
+  #       # ==================
+  #       # Note: It seems support for this type of feature was temporarly dropped
+  #       # in Polars 1.9 i.e. the data packing was improved but the avro
+  #       # serialization was not implemented. This code would fail with recent
+  #       # version of polars with: not yet implemented: write
+  #       # FixedSizeList(Field { name: "item", dtype: Float64,
+  #         is_nullable: true,
+  #       # metadata: {} }, 5) to avro.
+  #       "multi_f1": np.random.random(size=(n, 3)),
+  #       # # Labels
+  #       "label_class_binary1": np.random.choice([False, True], size=n),
+  #       "label_class_binary2": np.random.choice([0, 1], size=n),
+  #       "label_class_binary3": np.random.choice(["l1", "l2"], size=n),
+  #       "label_class_multi1": np.random.choice(["l1", "l2", "l3"], size=n),
+  #       "label_class_multi2": np.random.choice([0, 1, 2], size=n),
+  #       "label_regress1": np.random.random(size=n),
+  #   })
 
   def test_avro_from_raw_fastavro(self):
     tmp_dir = self.create_tempdir().full_path