+++ yggdrasil_decision_forests/port/python/ydf/model/model.cc
@@ -136,7 +136,7 @@ void init_model(py::module_& m) {
            py::arg("options"))
       // WARNING: This method releases the Global Interpreter Lock.
       .def("Save", WithStatus(&GenericCCModel::Save), py::arg("directory"),
-           py::arg("file_prefix"))
+           py::arg("file_prefix"), py::arg("pure_serving"))
       .def("Serialize", WithStatusOr(&GenericCCModel::Serialize))
       .def("name", &GenericCCModel::name)
       .def("task", &GenericCCModel::task)
@@ -233,6 +233,8 @@ void init_model(py::module_& m) {
            })
       .def("num_examples_per_tree",
            &IsolationForestCCModel::num_examples_per_tree)
+      .def("set_num_examples_per_tree",
+           &IsolationForestCCModel::set_num_examples_per_tree)
       .def_property_readonly_static("kRegisteredName",
                                     [](py::object /* self */) {
                                       return model::isolation_forest::