+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## HEAD
+
+### Feature
+
+-   Expose `model.save(..., pure_serving=True)` for saving a model without debug
+    information.
+-   Allow users to provide a training proto configuration to the learner.
+
 ## 0.9.0 - 2024-12-02
 
 ### Feature